test: move WebSocket close-code tests into a dedicated offline file
Consolidate the close-code and fragmented-close suites into
websocket-close-code.test.ts, which only uses a local raw-TCP server.
They were appended to websocket.test.js, whose pre-existing
ws.postman-echo.com tests hard-fail without internet access — keeping
the close-handshake coverage in its own network-free file makes it run
deterministically offline and groups the related cases together.
